fix decimals in movement summary

+++ app/controllers/charts_controller.rb
@@ -55,7 +55,7 @@ class ChartsController < ApplicationController
       data = PaperTrail::Version.all
     end
     data = data.where("'phase' = ANY (columns_changed)").select("phase_change, area").group_by(&:phase_change)
-    @data = Hash[data.map { |k,v| [k, [(v ? v.count : 0), v.sum { |v| v.area.to_f } ]] }]
+    @data = Hash[data.map { |k,v| [k, [(v ? v.count : 0), v.sum { |v| v.area.to_i } ]] }]
     @phases_in_order = %w(raw lamination inspection stock wip fg test nc scrap)
   end
 end
